refactor(go): move to different modules
